Added key to setVoltages functions for clarity   -Changed EIS messaging to match app expectation     -Changed EIS to testmode 3

+++ erdcWorkSpace/src/chemcrab_lib/craabUtil.c
@@ -489,9 +489,9 @@ int getVoltageInput(void){
 }
 
 //Returns a uint16_t representing  the UART input
-uint16_t getParameter(int dec){
+int getParameter(int dec){
   
-  uint16_t parameter = 0;
+  int parameter = 0;
   
   while(1){
     if(get_flag()){
